Added paru support

+++ install.sh
@@ -16,10 +16,10 @@ esac
 if [[ $start = "true" ]]; then
 
 
-echo "Updating system"
+echo -e "\nUpdating system\n"
 sudo pacman --noconfirm -Syu 
 
-echo "Installing base packages"
+echo -e "\nInstalling base packages\n"
 sudo pacman -S --noconfirm --needed base-devel git rsync
 
 echo "1) xf86-video-intel	2) xf86-video-amdgpu	3) none"
@@ -46,20 +46,36 @@ esac
 
 sudo pacman -S --noconfirm --needed $driver
 
-echo "Installing required packages from official repos"
+echo "\nInstalling required packages from official repos\n" && sleep 3
 sudo pacman -S --noconfirm --needed alacritty rxvt-unicode bspwm sxhkd i3-gaps obconf dunst picom xsettingsd lightdm-webkit2-greeter firefox nitrogen nautilus mousepad vim maim polkit-gnome network-manager-applet blueberry lxappearance xorg ttf-fira-code ttf-roboto noto-fonts noto-fonts-cjk noto-fonts-emoji npm nodejs rustup xdg-user-dirs
 
-echo "Installing yay"
-git clone https://aur.archlinux.org/yay.git ~/yay
-cd ~/yay
-makepkg -si
-rm -rf ~/yay
-cd ~/
+echo "1) yay	2) paru"
+read -r -p "What AUR helper would you like? (default 1) " helper
 
-echo "Installing required packages from AUR"
-yay -S --noconfirm --needed polybar betterlockscreen discord-canary zentile nerd-fonts-fira-code nerd-fonts-iosevka gtk3-nocsd-git
+if [[ $helper = "1" ]]; then
+	git clone https://aur.archlinux.org/yay.git ~/yay
+	cd ~/yay
+	makepkg -si
+	rm -rf ~/yay
+	cd ~/
 
-echo "Installing patched openbox"
+	aur="yay"
+fi
+
+if [[ $helper = "2" ]]; then
+	git clone https://aur.archlinux.org/paru.git ~/paru
+	cd ~/paru
+	makepkg -si
+	rm -rf ~/paru
+	cd ~/
+
+	aur="paru"
+fi
+
+echo -e "\nInstalling required packages from AUR\n" && sleep 3
+$aur -S --noconfirm --needed polybar betterlockscreen discord-canary zentile nerd-fonts-fira-code nerd-fonts-iosevka gtk3-nocsd-git
+
+echo -e "\nInstalling patched openbox\n" && sleep 3
 git clone https://github.com/Stardust-kyun/openbox ~/openbox
 cd ~/openbox
 ./bootstrap
@@ -69,21 +85,21 @@ sudo make install
 rm -rf ~/openbox
 cd ~/
 
-echo "Installing patched dmenu"
+echo -e "\nInstalling patched dmenu\n" && sleep 3
 git clone https://github.com/Stardust-kyun/dmenu ~/dmenu
 cd ~/dmenu
 sudo make clean install
 rm -rf ~/dmenu
 cd ~/
 
-echo "Installing powercord"
+echo -e "\nInstalling powercord\n" && sleep 3
 mkdir ~/.config
 git clone https://github.com/powercord-org/powercord ~/.config/powercord
 cd ~/.config/powercord
 npm i
 sudo npm run plug
 
-echo "Installing eww"
+echo -e "\nInstalling eww\n" && sleep 3
 git clone https://github.com/elkowar/eww ~/eww
 cd ~/eww
 cargo build --release
@@ -105,9 +121,8 @@ tar -xzf Arch.tar.gz
 tar -xzf Cabin.tar.gz
 tar -xzf NoelRed.tar.gz
 rm *.tar.gz
-# sudo sed -i '103 s/greeter-session = .*/greeter-session = lightdm-webkit2-greeter/' /etc/lightdm/lightdm.conf
+sudo sed -i 's/greeter-session = .*/greeter-session = lightdm-webkit2-greeter/' /etc/lightdm/lightdm.conf
 sudo systemctl enable lightdm
-# mv ~/.mozilla/firefox/default-release/chrome ~/.mozilla/firefox/*.default-release/
 xdg-user-dirs-update
 gsettings set org.gnome.nautilus.preferences always-use-location-entry true
